requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
tzdata>=2024.1
